ci: restructure and add run script
i've reorganized a lot of the code
src:
(read: "source") is typically where your impactful changes will go. It's the meat of the project.

dependencies:
this isn't following a convention, but it wasn't clear what was needed to get the project up and running. it's good to keep everything you need to get a repo going, within the repo. I saw from your build scripts, you were relying on a binary based on your C drive, but it meant there was no way for me to know exactly what's required.
this is a minor point, because if it's just you working on it, then it's not really a problem.
but in the event multiple people work on a project, or if you swap laptop or something, then you'll be thankful to have documented these requirements/dependencies

build:
if a file can be generated, it typically isn't committed to the repo. this includes binary files which don't make sense to keep a history of, as their history isn't comparable.
people do version their releases/builds, but it's not typically contained within the git history.
it's fine to do this for your sakes though, just wanted to add context

readme.md
it's good practice to leave yourself some instructions on how to get setup. this file is special, in that github will format it nicely, and it'll show up in their folder structure.

.gitignore
this is a special file used to exclude certain items from a repository. it's not very important for you to know about right now

run.bat
this shows how you can pass in the file you just built directly into the emulator, meaning you don't need to find it and run yourself, which was a little annoying